+++ resources/views/term/create.blade.php
@@ -10,7 +10,7 @@
             <x-laraknife.forms.combobox position="last" name="owner_id" label="Owner" :options="$optionsOwner" width2="4" />
             <x-laraknife.forms.string type="datetime-local" position="first" name="term" label="Term"
                 value="{{ $context->valueOf('term') }}" width2="4" />
-            <x-laraknife.forms.string position="last" name="duration" label="Duration" value="{{ $context->valueOf('duration') }}"
+            <x-laraknife.forms.string position="last" name="duration" label="Duration (minutes)" value="{{ $context->valueOf('duration') }}"
                 width2="4" />
             <x-laraknife.forms.string position="alone" name="title" label="Title" value="{{ $context->valueOf('title') }}"
                 width2="10" />