+++ templates/Http/Controllers/TermController.php
@@ -118,10 +118,11 @@ class TermController extends Controller
                 ];
             } else {
                 $conditions = [];
-                ViewHelper::addConditionComparism($conditions, $parameters, 'visibility_scope', 'visibility');
+                //ViewHelper::addConditionComparism($conditions, $parameters, 'visibility_scope', 'visibility');
                 ViewHelper::addConditionComparism($conditions, $parameters, 'owner_id', 'owner');
                 ViewHelper::addConditionPattern($conditions, $parameters, 'title,description', 'text');
                 ViewHelper::addConditionDateTimeRange($conditions, $parameters, 'from', 'to', 'term');
+                ViewHelper::addConditionVisible($conditions, $fields['visibility']);
                 $sql = DbHelper::addConditions($sql, $conditions);
             }
             $sql = DbHelper::addOrderBy($sql, $fields['_sortParams']);